feat(#170): 딥링크 관련 로직 구현 - MainNavigationGraph에 있던 로그인 시 이동 관련 로직을 MainActivity에서 설정 - MainActivity에서 Intent가 Null인지를 통해 아닐 경우 handleDeepLink함수에 intent를 넣어 인자 추출 및 적절한 위치로 이동 - 이를 위해 NavController를 MainScreen 쪽이 아니라 MainActivity로 수정
